+++ opentelemetry-proto/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Remove support for grpcio transport (#1534)
 - Add `schemars::JsonSchema` trait support with `with-schemars` feature (#1419)
 - Update protobuf definitions to v1.1.0 (#1154)